Files with statements only can be parsed
I don't know why the parser handles `ptIdentifier` so different here. But when you parse a file that starts with statements like the following then you receive an empty tree. 

A file like this:
```pascal
    MyObject.DoWhatever(5);
```

now results in a valid tree.
Was there any particular reason for handling `ptIdentifier` differently?